Merger auto-merge-rocm-systems-develop-729 into amd-npi-compute-external-merge-new-729 (#459)
* SWDEV-549518 - Enable logging dynamically through HIP APIS. (#1079)

* SWDEV-549518 - Enable logging dynamically through HIP APIS.

* SWDEV-549518 - Adding ROCProfiler related new API changes.

* rocprofiler-sdk changes for hip api additions.

---------


* [rocprof-compute] Pin ruff version for consistent formatting (#2680)

* pin ruff versions each to current latest

* Update rocprofiler-compute-formatting.yml

* Downgrade .pre-commit-config.yaml to match develop

* Fix Python Formatting (#2679)

Updated version of black to 26.1.0 updated some formatting rules


* Add automatic PyTorch library discovery for Python applications (#2623)

* Add automatic PyTorch library discovery for Python applications (#2623)

* Add notice for the newly deprecated env variables (#2690)

* [rocprofiler-compute] Use TheRock nightly builds in testing container (#2661)

* Use TheRock nightly builds in testing container

* Add HIP_DEVICE_LIB_PATH env var for hipcc to work

* Add HIP_PLATFORM env var for cmake hip package

* Add tarball placeholder

* Add -f to curl command to fail on HTTP error

* [rocprofiler-compute] Fix kernel/dispatch filtering (#2479)

* Fix kernel/dispatch fitlering in GUI

* Disallow --kernel and --dispatch filtering in analyze --gui mode since
  GUI frontend offers dropdown menu for kernel and dispatch filtering
    * Update CHANGELOG and documentation

* Gracefully handle N/A values

* Ensure workload path is valid before using it in GUI

* Ignore kernel filters if dispatch filters provided

* Add documentation for dispatch filtering overriding kernel filtering

* Fix typo

* Fix documentation

* remove unnecessary whitespace

* Address review comments

* Allow kernel/dispatch filtering with --gui

* Address review comments

* Address review comments

* Update CHANGELOG

* Fix formatting

* [rocprofiler-systems] Add build option for "examples" to specify gfx-arch (#2626)

## Motivation
 - Added `check_rocminfo` function that returns true if the provided regex was found, false otherwise. Can also use `GET_OUTPUT` to get the raw output filtered with or without a regex.
 - Moved `rocprofiler_systems_get_gfx_archs()` to `MacroUtilities.cmake` 
 - Added `rocprofiler_systems_lookup_gfx()`, which detects whether a given `gfx` is from the `instinct`, `radeon` or `apu` family.
 - Added `ROCPROFSYS_GFX_TARGETS` as a build argument. Used to specify the offloading architectures that GPU examples should compile for. If empty, defaults to whatever your system has.
 - GPU examples now check if the given `gfx` targets (from `ROCPROFSYS_GFX_TARGETS`) are supported.
 - OMPVV offload tests now only compile if `amdflang` version is `>= 20`
 - Improve link time by reducing the number of GFX targets that binaries need to support.
   - RCCL is now passed a `GPU_TARGETS` var specifying the architectures to build/link against.

* Reset HIP_VERSION_PATCH to 0 (#2590)

* [SWDEV-559965] Update Changelog for power cap type (#2647)

* [SWDEV-559965] Update Changelog for amd-smi set --power-cap

Updated Changelog to mention flexible argument
ordering for power cap type in amdsmi power cap set.
Corrected Changelog documentation on PPT1 reset
power_cap command.


* [rocprofiler-system]: Enable UCX Communication API tracing (#2306)

## Motivation

Enable UCX communication tracing and communication metadata 

## Technical Details

Implement UCX API wrappers to trace transport-layer communication. This adds communication data tracking and exposes “UCX Comm Send/Recv” timelines, enabling detailed analysis of MPI, OpenSHMEM, and other UCX-based runtime communication patterns.

- Implements function interception for UCX functions across multiple categories using gotcha component.
- Extended comm_data component to track UCX send/recv operations - Added ucx_send and ucx_recv labels for Perfetto counter tracks. Integrated UCX data tracking with existing MPI/RCCL tracking infrastructure.
- Added ROCPROFSYS_USE_UCX configuration option (enabled by default).
- Created FindUCX.cmake module for UCX header detection. Falls back to internal UCX headers if system headers not found.
- Updated all Dockerfiles  to include UCX dependencies.

* hip-issue-3876 : Take into account thread-local capture mode in checks for valid capture (#2177)

* Revert "rocr: Switch back to legacy IPC (#1744)" (#2676)

This reverts commit 7e4b62290c8b21506b32fccdb01213ad179c078f.

* SWDEV-558849 - Add support for static linking with ROCR (#2659)

* [rocprofiler-systems] Fix MPI recv_data calculation (#2694)

Fix incorrect `mpi_recv` calculation. It was using `_send_size` instead of `_recv_size` for `mpi_recv`.

* [ROCmInfo] docs: mono-repo changes and style edits (#2584)

* initial edits

* mono repo related updates

* standardize component name

* style edits

* more edits

* Update CODEOWNERS (#2705): add /project/amdsmi owner

* Use size_t datatype for global dimensions. (#2604)

* rocrtst: set HSA_ENABLE_INTERRUPT after TestExample creation (#2687)


* Rework clock based unit tests (#2646)

* Update hip_api_trace.cpp for HIP_RUNTIME_API_TABLE_STEP_VERSION 22

* Update hip_api_trace.hpp for HIP_RUNTIME_API_TABLE_STEP_VERSION 22

---------

+++ projects/clr/rocclr/platform/ndrange.hpp
@@ -202,9 +202,9 @@ struct HIPLaunchParams : public LaunchParams {
                   uint32_t globalX_remainder = 0, uint32_t globalY_remainder = 0,
                   uint32_t globalZ_remainder = 0, uint32_t clusterX = 1,
                   uint32_t clusterY = 1, uint32_t clusterZ = 1)
-                  : LaunchParams(static_cast<uint32_t>(gridX) * blockX + globalX_remainder,
-                                 static_cast<uint32_t>(gridY) * blockY + globalY_remainder,
-                                 static_cast<uint32_t>(gridZ) * blockZ + globalZ_remainder,
+                  : LaunchParams(static_cast<size_t>(gridX) * blockX + globalX_remainder,
+                                 static_cast<size_t>(gridY) * blockY + globalY_remainder,
+                                 static_cast<size_t>(gridZ) * blockZ + globalZ_remainder,
                                  blockX, blockY, blockZ, sharedMemBytes, clusterX, clusterY,
                                  clusterZ, gridX, gridY, gridZ, true /*hipParams*/) {}
 };
